test(chat): fold scroll-sync guard into existing adapter test, drop duplicate
The previous commit added tanstackMessageVirtualizer.test.ts, but a
tanstackMessageVirtualizer.test.tsx already existed (from #627). The two
same-base test files confused typescript-eslint's project service, failing
`npm run lint` in CI (the Test job) — a check I omitted locally before pushing.

Move the 'scrollToOffset dispatches a scroll event' guard into the existing
.tsx (adding scrollToOffset to its @tanstack mock) and delete the duplicate .ts.
Lint, typecheck, and the adapter tests (2) all pass.
